% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.mosaSTRS.R
\name{print.summary.mosaSTRS}
\alias{print.summary.mosaSTRS}
\title{Print a summary.mosaSTRS object}
\usage{
\method{print}{summary.mosaSTRS}(x, ...)
}
\arguments{
\item{x}{an object inheriting from class \code{summary.mosaSTRS},
representing the results of the function \code{mosallocSTRS()}}

\item{...}{some methods for this generic require additional arguments.
None are used in this method.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print-function for class \code{summary.mosaSTRS}.
}
